Changed some script things
Changed codegen.bat to not use hard-coded directories.
Added a script to set your PythonDir environment variable, so that you
can always run codegen.bat without having to provide it with an extra
argument or having to change the script every time a change is made to
it in the master branch.